# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Book Store
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /book:
    # our controller name
    x-swagger-router-controller: book
    get:
      description: get the books list
      operationId: getAllBooks
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetBookListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new book to the list
      operationId: saveBook
      # movie info to be stored
      parameters:
        - name: title
          description: Book properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /book/{id}:
    # our controller name
    x-swagger-router-controller: book
    get:
      description: get a book
      operationId: getBook
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetBookResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a book
      operationId: updateBook
      # define the parameters
      parameters:
        - name: id
          description: book id
          type: string
          in: path
          required: true
        - name: title
          description: book properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    delete:
      description: delete a book
      operationId: deleteBook
      # define the parameters
      parameters:
        - name: id
          description: book id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GetBookListResponse:
    required:
      - books
    properties:
      # The array of books
      books:
        type: array
        items: 
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            year:
              type: number
            author:
              type: string
              
              
  GetBookResponse:
    required:
      - _id
      - title
      - year
      - author
    properties:
      _id:
        type: string
      title: 
        type: string
      year:
        type: number
      author:
        type: string
  
  Book:
    type: object
    properties:
      title:
        type: string
        description: task object name
      year:
        type: number
        description: task description
      author:
        type: string
        description: Author name
    required:
      - title
      - year
      - author
      
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
      
  
      
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        